feat(prejoin): Move startConference logic to conference middleware.
