+++ react/features/base/conference/middleware.web.ts
@@ -8,11 +8,13 @@ import { isPrejoinPageVisible } from '../../prejoin/functions';
 import { iAmVisitor } from '../../visitors/functions';
 import { CONNECTION_DISCONNECTED, CONNECTION_ESTABLISHED } from '../connection/actionTypes';
 import { hangup } from '../connection/actions.web';
-import { JitsiConferenceErrors } from '../lib-jitsi-meet';
+import { JitsiConferenceErrors, browser } from '../lib-jitsi-meet';
 import { gumPending, setInitialGUMPromise } from '../media/actions';
 import { MEDIA_TYPE } from '../media/constants';
 import { IGUMPendingState } from '../media/types';
 import MiddlewareRegistry from '../redux/MiddlewareRegistry';
+import { replaceLocalTrack } from '../tracks/actions.any';
+import { getLocalTracks } from '../tracks/functions.any';
 
 import {
     CONFERENCE_FAILED,
@@ -147,10 +149,46 @@ MiddlewareRegistry.register(store => next => action => {
         break;
     }
     case CONNECTION_ESTABLISHED: {
-        const state = getState();
+        if (isPrejoinPageVisible(getState())) {
+            let { initialGUMPromise } = getState()['features/base/media'];
 
-        if (!isPrejoinPageVisible(state)) {
-            const { initialGUMPromise = Promise.resolve({ tracks: [] }) } = state['features/base/media'];
+            initialGUMPromise = initialGUMPromise || Promise.resolve({ tracks: [] });
+
+            initialGUMPromise.then(() => {
+                const state = getState();
+                let localTracks = getLocalTracks(state['features/base/tracks']);
+                const trackReplacePromises = [];
+
+                // Do not signal audio/video tracks if the user joins muted.
+                for (const track of localTracks) {
+                    // Always add the audio track on Safari because of a known issue where audio playout doesn't happen
+                    // if the user joins audio and video muted.
+                    if ((track.muted && !(browser.isWebKitBased() && track.jitsiTrack
+                            && track.jitsiTrack.getType() === MEDIA_TYPE.AUDIO)) || iAmVisitor(state)) {
+                        trackReplacePromises.push(dispatch(replaceLocalTrack(track.jitsiTrack, null))
+                            .catch((error: any) => {
+                                logger.error(`Failed to replace local track (${track.jitsiTrack}) with null: ${error}`);
+                            }));
+                    }
+                }
+
+                Promise.allSettled(trackReplacePromises).then(() => {
+
+                    // Re-fetch the local tracks after muted tracks have been removed above.
+                    // This is needed, because the tracks are effectively disposed by the replaceLocalTrack and should
+                    // not be used anymore.
+                    localTracks = getLocalTracks(getState()['features/base/tracks']);
+
+                    const jitsiTracks = localTracks.map((t: any) => t.jitsiTrack);
+
+
+                    return APP.conference.startConference(jitsiTracks);
+                });
+            });
+        } else {
+            let { initialGUMPromise } = getState()['features/base/media'];
+
+            initialGUMPromise = initialGUMPromise || Promise.resolve({ tracks: [] });
 
             initialGUMPromise.then(({ tracks }) => {
                 let tracksToUse = tracks ?? [];
